Move schedule manager in its own file
